+++ client/src/components/GlobalComponents/landingPage/MainLanding.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import FrontDeskLanding from '../frontDesk/FrontDeskLanding';
-import LandingButtons from '../landingPage/MainLandingButtons';
-import useChoosePath from '../landingPage/useChoosePath';
-import HMLanding from '../maintenance/HMLanding';
-import ManagementLanding from '../management/ManagementLanding';
-import FlexCenterContainer from './FlexCenterContainer';
-import TimeSheet from '../GlobalComponents/timeSheet/TimeSheetContainer';
+import FrontDeskLanding from '../../frontDesk/FrontDeskLanding';
+import LandingButtons from './MainLandingButtons';
+import useChoosePath from './useChoosePath';
+import HMLanding from '../../maintenance/HMLanding';
+import ManagementLanding from '../../management/ManagementLanding';
+import FlexCenterContainer from '../../styledElements/FlexCenterContainer';
+import TimeSheet from '../timeSheet/TimeSheetContainer';
 import { MainProvider } from './MainContext';
 import { useAuth0 } from '@auth0/auth0-react';
