+++ client/src/components/GlobalComponents/landingPage/useChoosePath.jsx
