+++ client/src/components/GlobalComponents/login/Login.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { colors } from '../styledElements/styleGuid';
+import { colors } from '../../styledElements/styleGuid';
 import LoginButton from './LoginButton';
 
 const Wrapper = styled.main`
