+++ client/src/components/GlobalComponents/login/LogoutButton.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
-import FormButton from '../styledElements/FormButton';
+import FormButton from '../../styledElements/FormButton';
 
-import {REACT_APP_REDIRECT_URL} from '../../../../env/config';
+import {REACT_APP_REDIRECT_URL} from '../../../../../env/config';
 
 const LogoutButton = (authenticated) => {
   const { logout } = useAuth0();
