+++ client/src/components/GlobalComponents/timeSheet/TimeSheetContainer.jsx
@@ -1,10 +1,18 @@
 import React from 'react';
+
+// styled Component
 import HalfRoundDiv from '../../styledElements/HalfRoundDiv';
 import Title from '../../styledElements/ModalTitle';
+import FlexCenterContainer from '../../styledElements/FlexCenterContainer';
+
+// React Components that have real functionality
 import TimeSheetFormList from './TimeSheetFormList';
 import BackArrow from '../../styledElements/BackArrow';
-import FlexCenterContainer from '../../landingPage/FlexCenterContainer';
 
+/**
+ * This is the main container for the time sheet widget. Only the BackArrow and TimeSheetFormList
+ * @param {Function} back - This is the function that allows the user to close this form passed as props from the main landing
+ */
 const TimeSheetContainer = ({ back }) => (
   <div>
     <BackArrow margin={'40px 40px'} onClick={back} location={'Main Homepage'} />
