+++ client/src/components/GlobalComponents/timeSheet/timeSheetTemplate.js
@@ -1,3 +1,5 @@
+// this is designed this way be we need both the format the db/server expects and a nice name version for display
+// we have to use and array of objects over just and object bc JS objects are not ordered
 const timeSheetTemplate = [
   { day: 'monday', name: 'Monday' },
   { day: 'tuesday', name: 'Tuesday' },
