+++ client/src/components/frontDesk/CreateBooking/AvailableRooms.jsx
@@ -1,13 +1,12 @@
-import React from "react";
-import HalfRoundDiv from "../../styledElements/HalfRoundDiv.jsx";
+import React from 'react';
+import HalfRoundDiv from '../../styledElements/HalfRoundDiv.jsx';
 import ListEntry from './ListEntry.jsx';
-import FlexCenterContainer from '../../landingPage/FlexCenterContainer.jsx';
 
 let AvailableRooms = (props) => {
   return (
     <HalfRoundDiv id={'roomListContainer'}>
-      <h1 className="title">Available Room Types</h1>
-      <div className="roomList">
+      <h1 className='title'>Available Room Types</h1>
+      <div className='roomList'>
         {props.availableRooms.map((room) => {
           return (
             <ListEntry
@@ -16,7 +15,9 @@ let AvailableRooms = (props) => {
                 props.selectRoom(event);
               }}
               table={room}
-            />)})}
+            />
+          );
+        })}
       </div>
     </HalfRoundDiv>
   );
