+++ client/src/components/frontDesk/CreateBooking/ConfirmationPage.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import FlexCenterContainer from '../../landingPage/FlexCenterContainer.jsx';
 import FormButton from '../../styledElements/FormButton.jsx';
 
 let ConfirmationPage = (props) => {
   return (
-    <div className="confirmation">
-      <h1 className="title">Confirmation Page</h1>
-      <ul className="infoList">
+    <div className='confirmation'>
+      <h1 className='title'>Confirmation Page</h1>
+      <ul className='infoList'>
         <li>{`First Name: ${props.firstName}`}</li>
         <li>{`Last Name: ${props.lastName}`}</li>
         <li>{`Phone Number: ${props.phone}`}</li>
@@ -15,16 +14,17 @@ let ConfirmationPage = (props) => {
         <li>{`Check In Date: ${props.checkIn}`}</li>
         <li>{`Check Out Date: ${props.checkOut}`}</li>
       </ul>
-      <FormButton onClick={() => {
-        props.submitBooking();
-        props.clickBack();
-      }}
-      id={'submitConfirm'}
+      <FormButton
+        onClick={() => {
+          props.submitBooking();
+          props.clickBack();
+        }}
+        id={'submitConfirm'}
       >
         Confirm Booking
       </FormButton>
     </div>
-  )
-}
+  );
+};
 
 export default ConfirmationPage;
