+++ client/src/components/frontDesk/FrontDeskLanding.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import LandingButtons from '../landingPage/LandingButtons';
-import useChoosePath from '../landingPage/useChoosePath';
+import LandingButtons from '../GlobalComponents/landingPage/LandingButtons';
+import useChoosePath from '../GlobalComponents/landingPage/useChoosePath';
 import BackArrow from '../styledElements/BackArrow';
 import CreateBookingForm from './CreateBooking/CreateBookingForm.jsx';
 import ListMaster from '../GlobalComponents/ListMaster';
-import FlexCenterContainer from '../landingPage/FlexCenterContainer';
+import FlexCenterContainer from '../styledElements/FlexCenterContainer';
 import ViewAddEditRoom from '../management/ViewAddEditRoom.jsx';
 import CheckoutLanding from './checkout/CheckoutLanding';
 import SearchForReservations from './checkIn/01searchForReservations';
@@ -32,7 +32,7 @@ const FrontDeskLanding = ({ back }) => {
     },
     { title: 'Check-Out', onClick: () => setPaths('checkOut') },
     { title: 'View Rooms', onClick: () => setPaths('viewRooms') },
-    { title: 'Create Booking', onClick: () => setPaths('createBooking') }
+    { title: 'Create Booking', onClick: () => setPaths('createBooking') },
   ];
 
   // return to main landing
@@ -66,7 +66,7 @@ const FrontDeskLanding = ({ back }) => {
           />
         ) : paths.createBooking ? (
           <FlexCenterContainer>
-            <CreateBookingForm clickBack={clickBack}/>
+            <CreateBookingForm clickBack={clickBack} />
           </FlexCenterContainer>
         ) : (
           <FlexCenterContainer>
