+++ client/src/components/maintenance/HMLanding.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useContext } from 'react';
-import LandingButtons from '../landingPage/LandingButtons';
-import useChoosePath from '../landingPage/useChoosePath';
+import LandingButtons from '../GlobalComponents/landingPage/LandingButtons';
+import useChoosePath from '../GlobalComponents/landingPage/useChoosePath';
 import BackArrow from '../styledElements/BackArrow';
 import ListMaster from '../GlobalComponents/ListMaster';
 import AddTaskForm from './AddTaskForm.jsx';
-import FlexCenterContainer from '../landingPage/FlexCenterContainer';
+import FlexCenterContainer from '../styledElements/FlexCenterContainer';
 import ViewAddEditRoom from '../management/ViewAddEditRoom.jsx';
-import { MainContext } from '../landingPage/MainContext.jsx';
+import { MainContext } from '../GlobalComponents/landingPage/MainContext.jsx';
 
 const HMLanding = ({ back }) => {
   // set state to toggle for conditional rendering
@@ -17,7 +17,7 @@ const HMLanding = ({ back }) => {
     landing: true,
   });
 
-  const {id, name} = useContext(MainContext);
+  const { id, name } = useContext(MainContext);
 
   //set up for transition to room list page
   const [backColor, setBackColor] = useState('white');
@@ -68,7 +68,7 @@ const HMLanding = ({ back }) => {
         ) : paths.addTask ? (
           //  add task component
           <FlexCenterContainer>
-            <AddTaskForm employeeId={id} name={name} clickBack={clickBack}/>
+            <AddTaskForm employeeId={id} name={name} clickBack={clickBack} />
           </FlexCenterContainer>
         ) : (
           <FlexCenterContainer>
