+++ package.json
@@ -25,26 +25,26 @@
     "sass-loader": "^10.1.1",
     "style-loader": "^2.0.0",
     "webpack": "^5.13.0",
+    "nodemon": "^2.0.7",
+    "path": "^0.12.7",
+    "express": "^4.17.1",
+    "eslint": "^7.17.0",
+    "eslint-plugin-react": "^7.22.0",
+    "enzyme": "^3.11.0",
+    "sass": "^1.32.7",
+    "node-sass": "^5.0.0",
+    "jest": "^26.6.3",
     "webpack-cli": "^4.3.1"
   },
   "dependencies": {
     "@auth0/auth0-react": "^1.3.0",
     "auth0-react": "^3.1.0",
     "axios": "^0.21.1",
-    "enzyme": "^3.11.0",
-    "eslint": "^7.17.0",
-    "eslint-plugin-react": "^7.22.0",
-    "express": "^4.17.1",
-    "jest": "^26.6.3",
-    "node-sass": "^5.0.0",
-    "nodemon": "^2.0.7",
-    "path": "^0.12.7",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-modal": "^3.12.1",
     "react-router-dom": "^5.2.0",
     "react-spinners": "^0.10.6",
-    "sass": "^1.32.7",
     "styled-components": "^5.2.1"
   }
 }
